IPv6 support and various bug fixes

+++ entrypoint.sh
@@ -7,15 +7,11 @@ MASTER_IFACE="eth0"
 # Some distros, like k3s, symlink CNIs to the same binary. Trying to copy files like this will
 # mean we corrupt the binaries
 for filename in /cni/*; do
-  rm /host/cni_bin/$(basename $filename)
+  [ -e /host/cni_bin/$(basename $filename) ] && rm /host/cni_bin/$(basename $filename)
   cp $filename /host/cni_bin/$(basename $filename)
 done
 
 sed "s/MASTER_IFACE/$MASTER_IFACE/g" bridge-cni.tmpl > /host/cni_net/${PRIORITY:-10}-bridge.conflist
-
-/sbin/ip link add mac0 link eth0 type macvlan mode bridge || true
-/sbin/ip addr add $LOCAL_HOST_CIDR dev mac0 || true
-/sbin/ip link set mac0 up || true
 
 rm -f /run/cni/dhcp.sock
